+++ backend/__init__.py
@@ -1,8 +1,12 @@
 
 
 version="v2"
+version3="v3"
 
 version_prefix=f"/api/{version}"
+
+version3_prefix=f"/api/{version3}"
+